data_pipeline/requirements.txt
pandas
psycopg2-binary
xlrd
jupyter
fastapi
uvicorn
pydantic
SQLAlchemy
python-dotenv
requests
